+++ rpv3_options.h
@@ -12,10 +12,10 @@ extern "C" {
 
 /* Version information */
 /* Version information */
-#define RPV3_VERSION "1.3.1"
+#define RPV3_VERSION "1.3.2"
 #define RPV3_VERSION_MAJOR 1
 #define RPV3_VERSION_MINOR 3
-#define RPV3_VERSION_PATCH 1
+#define RPV3_VERSION_PATCH 2
 
 
 /* Return codes */